% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_corpus_positions.R
\name{add_corpus_positions}
\alias{add_corpus_positions}
\title{Add Corpus Positions to Metadata Table.}
\usage{
add_corpus_positions(x, verbose = TRUE)
}
\arguments{
\item{x}{a \code{corpusdata} object}

\item{verbose}{logical, whether to be talkative}
}
\description{
Add columns cpos_left and cpos_right to table with structural attributes. A
precondition is that a column 'id' is present in tables 'tokenstream' and 'metadata'."
}
