% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner.R
\name{as.Annotation}
\alias{as.Annotation}
\title{Coerce III}
\usage{
as.Annotation(x, type = c("word", "sentence"))
}
\description{
Coerce III
}
\examples{
library(polmineR)
use("GermaParl")
use("UNGA")
p <- partition("KEYWORDS_EN", text_date = "2012-12-31")
s <- NLP::String(as.String(p))
a <- as.Annotation(p, type = "word")
sent <- as.Annotation(p, type = "sentence")
a1 <- c(sent, a)
foo <- s[a1[a1$type == "word"]]

sentences <- s[a1[a1$type == "sentence"]]

library(openNLP)
install.packages("openNLPmodels.de", repos = "http://datacube.wu.ac.at/", type = "source")

entity_annotator <- Maxent_Entity_Annotator(language = "de")
ner <- annotate(s, entity_annotator, a1)
s[ner]
s[entity_annotator(s, a2)]
}
