% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_corpusdata.R
\name{as_corpusdata}
\alias{as_corpusdata}
\title{Parse an XML/TEI file as data.table}
\usage{
as_corpusdata(x, body = "//body", meta = NULL, mc = NULL)
}
\arguments{
\item{x}{a single filename, a character vector of filenames, or a directory with XML files}

\item{body}{an xpath expression defining the body of the xml document}

\item{meta}{a named character vector with xpath expressions}

\item{mc}{a numeric/integer value, number of cores to use}
}
\description{
The procedure is inspired by the tidytext paradigm, but
uses a data.table for efficiency.
}
