% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwb.R
\name{cwb_make}
\alias{cwb_make}
\title{Run cwb-make Command}
\usage{
cwb_make(corpus, registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  verbose = TRUE)
}
\arguments{
\item{corpus}{the CWB corpus}

\item{registry_dir}{the registry directory}

\item{verbose}{logical}
}
\description{
Run cwb-make Command
}
