% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cwb_corpus_dir}
\alias{cwb_corpus_dir}
\alias{cwb_registry_dir}
\alias{cwb_directories}
\alias{cwb_registry_dir_create}
\alias{cwb_corpus_dir_create}
\title{Guess corpus directory with data directories for corpora}
\usage{
cwb_corpus_dir(registry_dir)

cwb_registry_dir()

cwb_directories(prefix = "~/cwb", create = FALSE, ask = interactive())

cwb_registry_dir_create(prefix = "~/cwb", ask = interactive())

cwb_corpus_dir_create(prefix = "~/cwb", ask = interactive())
}
\arguments{
\item{registry_dir}{Path to the directory with registry files. If missing, an
internal heuristic will try to get the value of environment variable
CORPUS_REGISTRY.}

\item{prefix}{The base path that will be prefixed}

\item{create}{A \code{logical} value, whether directories shall be created.}

\item{ask}{A \code{logical} value, whether to prompt user before creating
directories.}
}
\value{
Length-one \code{character} vector with path of suggested data directory, or
  \code{NULL} if the data directory cannot be plausible guessed.
}
\description{
Apply heuristic to make a plausible suggestion for a corpus directory where
data directories for corpora reside.
}
\details{
First, directories in the parent directory of the registry directory that
contain "corpus" or "corpora" are suggested. If this does not yield a result,
the home directories stated in the registry files are evaluated. If there is
one unique parent directory of data directories (after removing temporary
directories and directories within packages), this directory is suggested.
}
