% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_corpusdata}
\alias{encode_corpusdata}
\alias{encode_structural_attribute}
\alias{encode_positional_attribute}
\title{Encode CWB corpus.}
\usage{
encode_corpusdata(x, corpus, p_attributes = "word", s_attributes = NULL,
  encoding, registry_dir = Sys.getenv("CORPUS_REGISTRY"), data_dir = NULL,
  verbose = TRUE)

encode_structural_attribute(x, s_attribute, data_dir, encoding,
  verbose = TRUE)

encode_positional_attribute(x, p_attribute = "word", corpus, encoding,
  registry_dir, data_dir, make = TRUE, verbose = TRUE)
}
\arguments{
\item{corpus}{corpus name of the CWB corpus}

\item{p_attributes}{columns of .Object with tokens (such as word/pos/lemma)}

\item{s_attributes}{columns of .Object that will be encoded as structural attributes}

\item{encoding}{encoding/charst of the CWB corpus}

\item{registry_dir}{corpus registry, the directory where registry files are stored}

\item{data_dir}{directory with indexed corpus files}

\item{verbose}{logical, whether to be verbose}

\item{s_attribute}{a single s-attribute}

\item{.Object}{input object}

\item{corpus}{the name of the CWB corpus}

\item{data_dir}{directory where to create directory for indexed corpus files}
}
\description{
\code{encode_structural_attribute} will add a structural attribute to a corpus 
from a data.table with three columns: The left corpus position, the right corpus position and the value of a s-attribute
that will be encoded.
}
