% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cwb.R
\name{install_cwb}
\alias{install_cwb}
\title{Install Corpus Workbench}
\usage{
install_cwb()
}
\description{
Install precompiled version of the Corpus Workbench in the directory
extdata/cwb/CWB. This is a precondition to offer full CQP functionality
on Windows systems.
}
