% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner.R
\name{partition_to_Annotation}
\alias{partition_to_Annotation}
\title{Coerce II}
\description{
Coerce II
}
\examples{
library(polmineR)
use("UNGA")
p <- partition("UNGA", date = "2018-03-20", who = "Rybakov")
p <- partition("UNGA", date = "2018-03-20", who = "Al Hussein")
p <- partition("UNGA", date = "2018-03-20")
s <- as(p, "String")
a <- as(p, "Annotation")

s[a[a$type == "sentence"]] # show sentences

library(openNLP)
install.packages("openNLPmodels.de", repos = "http://datacube.wu.ac.at/", type = "source")

entity_annotator <- Maxent_Entity_Annotator(language = "en", kind = "person")
entity_annotator <- Maxent_Entity_Annotator(language = "en", kind = "date")
entity_annotator <- Maxent_Entity_Annotator(language = "en", kind = "location")
entity_annotator <- Maxent_Entity_Annotator(language = "en", kind = "organization")
ner <- NLP::annotate(s, entity_annotator, a)
s[ner[ner$type == "entity"]]
}
