% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_corpus_registry_envvar}
\alias{use_corpus_registry_envvar}
\title{Set CORPUS_REGISTRY environment variable}
\usage{
use_corpus_registry_envvar(path)
}
\arguments{
\item{path}{Path of the registry directory.}
}
\description{
Set CORPUS_REGISTRY environment variable
}
