% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{get_zenodo_record_metadata}
\alias{get_zenodo_record_metadata}
\alias{as_bibentry}
\alias{as_bibentry.zenodo_record_metadata}
\title{Get and process information from Zenodo.}
\usage{
get_zenodo_record_metadata(doi, timeout.ms = 5000)

as_bibentry(x)

\method{as_bibentry}{zenodo_record_metadata}(x)
}
\arguments{
\item{doi}{The Document Object Identifier (doi) pointing to a Zenodo record.}

\item{timeout.ms}{Time allowed for the Zenodo API to respond (in
milliseconds). Defaults to 5000 (milliseconds).}

\item{x}{An object of class \code{zenodo_record_metadata}.}
}
\description{
Zenodo is an open science repository hosted by CERN which can be readily be
used to deposit corpora (see \href{https://zenodo.org/}{Zenodo website}). The
'cwbtools' package includes basic functionality to retrieve and process the
metadata of Zenodo records.
}
\details{
The function \code{get_zenodo_record_metadata} will retrieve
  information from the Zenodo API for the record identified via a DOI
  (argument \code{doi}). The return value is an object of class
  \code{zenodo_record_metadata} (S3 class).

A method \code{as_bibentry} is available for the class
  \code{zenodo_record_metadata}. The resulting \code{bibentry}-class object
  can be turned into a recommended citation easily using the \code{format} or
  \code{print} methods for \code{bibentry} class objects (see
  \link[utils]{bibentry}).
}
\examples{
\dontrun{
gps_meta <- get_zenodo_record_metadata(doi = "https://doi.org/10.5281/zenodo.3823245")
as_bibentry(gps_meta)
}
}
