% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zenodo_get_tarball}
\alias{zenodo_get_tarball}
\title{Download corpus tarball from Zenodo}
\usage{
zenodo_get_tarball(
  url,
  destfile = tempfile(fileext = ".tar.gz"),
  verbose = TRUE,
  progress = TRUE
)
}
\value{
Filename of downloaded corpus, to serve as input (argument ``)
}
\description{
Download corpus tarball from Zenodo. Both freely available data and
restricted access is supported.
}
\examples{
fname <- zenodo_get_tarball(url = "https://zenodo.org/record/3823245")
corpus_install(fname)
}
