% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{as.Annotation}
\alias{as.Annotation}
\alias{decode}
\title{Decode Corpus as Annotation Object.}
\usage{
decode(x, type = c("word", "sentence"))
}
\arguments{
\item{x}{An object that can be processed by \code{polmineR::get_token_stream}.}

\item{type}{Either "word" or "sentence".}
}
\value{
An \code{Annotation} object.
}
\description{
Decode Corpus as Annotation Object.
}
\examples{
\dontrun{
use("RcppCWB")
p <- partition("UNGA", date = "2018-03-20", who = "Al Hussein")
s <- NLP::String(as.String(p))
a <- as.Annotation(p, type = "word")
sent <- as.Annotation(p, type = "sentence")
a1 <- c(sent, a)

words <- s[a1[a1$type == "word"]]
sentences <- s[a1[a1$type == "sentence"]]
}
}
