% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner.R
\name{conll_get_regions}
\alias{conll_get_regions}
\title{Extract regions from NER annotations (CoNNL format).}
\usage{
conll_get_regions(x)
}
\description{
Extract regions from NER annotations (CoNNL format).
}
\examples{
x <- data.frame(
  token = c("Die", "Bundeskanzlerin", "Angela", "Merkel", "hält", "im", "Bundestag", "eine", "Rede", "."),
  ne = c("O", "O", "B-PERS", "I-PERS", "O", "O", "B-ORG", "O", "O", "O"),
  stringsAsFactors = FALSE
)
x[["cpos"]] <- 100L:(100L + nrow(x) - 1L)
tab <- conll_get_regions(x)
}
