% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{install.corpus}
\alias{install.corpus}
\alias{corpus_install}
\alias{corpus_packages}
\alias{corpus_copy}
\alias{corpus_rename}
\alias{corpus_remove}
\alias{corpus_as_tarball}
\title{Install and Manage corpora.}
\usage{
corpus_install(pkg = NULL, repo = "http://polmine.sowi.uni-due.de/packages",
  tarball = NULL, lib = .libPaths()[1], verbose = TRUE, user = NULL,
  password = NULL, ...)

corpus_packages()

corpus_copy(old, new, registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  verbose = TRUE)

corpus_rename(old, new, registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  verbose = TRUE)

corpus_remove(corpus, registry_dir = Sys.getenv("CORPUS_REGISTRY"))

corpus_as_tarball(corpus, registry_dir, tarfile, verbose = TRUE)
}
\arguments{
\item{pkg}{data package name}

\item{repo}{URL of the repository}

\item{tarball}{the URL or local path to a tarball with a CWB indexed corpus}

\item{lib}{directory for R packages, defaults to \code{.libPaths()[1]}; the path may not 
include a whitespace sign}

\item{verbose}{logical, whether to be verbose}

\item{user}{A user name that can be specified to download a corpus from a password protected site.}

\item{password}{A password that can be specified to download a corpus from a password protected site.}

\item{...}{Further parameters that will be passed into
\code{install.packages}, if argument \code{tarball} is \code{NULL}, or into
or \code{download.file}, if \code{tarball} is specified.}

\item{old}{name of the (old) corpus}

\item{new}{name of the (new) corpus}

\item{registry_dir}{directory of registry}

\item{corpus}{a CWB corpus}

\item{tarfile}{Filename of tarball.}
}
\description{
Utitlity functions for making the installation of indexed CWB corpora wrapped
into R data packages as easy as possible. A data package with a CWB corpus is
assumed to include a directory \code{/extdata/cwb/registry} for registry
files and a directory \code{/extdata/cwb/indexed_corpora} for the inexed
corpus files.
}
\details{
The \code{corpus_install} function combines two steps necessary to
  install a CWB corpus. First, it calls
  \code{install.packages}, then it resets the path pointing to the directory
  with the indexed corpus files in the registry file. The package will be
  installed to the standard library directory for installing R packages
  (\code{.libPaths()[1]}). Another location can be used by stating the param
  'lib' explicitly (see documentation for \code{\link{install.packages}}).
  The function can also be used to install a corpus from a password protected
  repository. Further parameters are handed over to install.packages, so you
  might add \code{method = "wget" extra = "--user donald --password duck"}.
  See examples how to check whether the directory has been set correctly.

\code{corpus_packages} will detect the packages that include CWB
  corpora.

\code{corpus_copy} will create a copy of a corpus (useful for
  experimental modifications, for instance).

\code{corpus_rename} will rename a corpus, affecting the name of the
  registry file, the corpus id, and the name of the directory where data
  files reside.

\code{corpus_remove} can be used to drop a corpus.

\code{corpus_as_tarball} will create a tarball (.tar.gz-file) with
  two subdirectories. The 'registry' subdirectory will host the registry file
  for the tarred corpus. The data files will be put in a subdirectory with
  the corpus name in the 'indexed_corpora' subdirectory.
}
\seealso{
For managing registry files, see \code{\link{registry_file_parse}}
for switching to a packaged corpus.
}
