% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_file.R
\name{registry_file_parse}
\alias{registry_file_parse}
\alias{registry_file_compose}
\alias{registry_data}
\alias{registry_file_write}
\title{Parse and Compose Registry Files.}
\usage{
registry_file_parse(corpus, registry_dir = Sys.getenv("CORPUS_REGISTRY"))

registry_file_compose(x)

registry_data(name, id, home, info = file.path(home, ".info"),
  properties = c(charset = "utf-8"), p_attributes,
  s_attributes = character())

registry_file_write(data, corpus,
  registry_dir = Sys.getenv("CORPUS_REGISTRY"), ...)
}
\arguments{
\item{corpus}{a CWB corpus}

\item{registry_dir}{directory with registry files}

\item{x}{an object of class \code{registry_data}}

\item{name}{long descriptive name of corpus}

\item{id}{short name of corpus}

\item{home}{data directory for indexed corpus}

\item{info}{filename of info file}

\item{properties}{named character vector with corpus properties, should include 'charset'}

\item{p_attributes}{positional attributes to declare}

\item{s_attributes}{structural attributes to declare}

\item{data}{registry_data}

\item{...}{further parameters}
}
\description{
A set of functions to parse, create and write registry files.
}
\details{
\code{registry_file_parse} will return an object of class \code{registry_data}.

See the appendix to the 'Corpus Encoding Tutorial' (http://cwb.sourceforge.net/files/CWB_Encoding_Tutorial.pdf)
includes an explanation of the registry file format.
}
\examples{
regdata <- registry_file_parse(
  corpus = "REUTERS",
  registry_dir = system.file(package = "RcppCWB", "extdata", "cwb", "registry")
  )
}
