% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_attribute.R
\name{s_attribute_encode}
\alias{s_attribute_encode}
\title{Encode Structural Attribute.}
\usage{
s_attribute_encode(values, data_dir, s_attribute, corpus, region_matrix,
  method = c("R", "CWB"), registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  encoding, verbose = TRUE)
}
\arguments{
\item{values}{a character vector with the values of the structural attribute}

\item{data_dir}{the data directory where to write the files}

\item{s_attribute}{atomic character vector, the name of the structural attribute}

\item{corpus}{the CWB corpus}

\item{region_matrix}{two-column matrix with corpus positions}

\item{method}{either 'R' or 'CWB'}

\item{registry_dir}{registry directory}

\item{encoding}{encoding of the data}

\item{verbose}{logicalds}
}
\description{
Pure R implementation to generate structural attributes.
}
