% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{zenodo_download_corpus}
\alias{zenodo_download_corpus}
\alias{zenodo_info}
\title{Download corpus from Zenodo}
\usage{
zenodo_download_corpus(
  corpus = "UNGAMINI",
  doi = "https://doi.org/10.5281/zenodo.3748858",
  registry_dir = cwbtools::cwb_registry_dir(),
  quiet = FALSE,
  ask = interactive()
)

zenodo_info(doi)
}
\arguments{
\item{corpus}{ID of corpus to be downloaded.}

\item{doi}{The DOI (Digital Object Identifier) of the corpus tarball at
zenodo, presented as a hyperlink. Defaults to the DOI of a small sample
corpus (UNGAMINI).}

\item{registry_dir}{Path to the system registry directory. Defaults to value
of \code{cwbtools::cwb_registry_dir()}.}

\item{quiet}{Whether to suppress progress messages, defaults to \code{FALSE}.}

\item{ask}{A \code{logical} value, whether to ask for user input before
replacing an existing corpus.}
}
\value{
A logical value, \code{TRUE} if the corpus has been installed
  successfully.
}
\description{
Zenodo is an open science repository hosted by CERN.

\code{zenodo_info} is a helper to extract
}
